pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["Rauan Mangitov <rmangitov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
